% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotated_hymns}
\alias{annotated_hymns}
\title{Hymns, tokenized and POS-tagged}
\format{
A tibble with 153,717 rows and 7 variables:
\describe{
\item{doc_id}{Official hymn number}
\item{paragraph_id}{Line number in hymn}
\item{token_id}{Token number in line}
\item{token}{The original token, i.e. word}
\item{lemma}{The lemmatized, i.e. dictionary form, of token}
\item{upos}{POS-tag, i.e. part-of-speach, like VERB or PUNCT for punctuation}
\item{vowels}{Number of vowels in token -- good for finding alternative words
using the \code{cut_up()} function}
}
}
\usage{
annotated_hymns
}
\description{
The \code{hymn} dataset, tokenized and tagged with part-of-speach (POS).
Each line represent a word from a given hymn.
}
\keyword{datasets}
