% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_annotation.R
\name{collapse_annotation}
\alias{collapse_annotation}
\title{'Collapses' a annotation back into text}
\usage{
collapse_annotation(df)
}
\arguments{
\item{df}{A tibble based on output from \code{udpipe::udpipe_annotate()}
which is assumed to contain \code{token}, \code{doc_id},
and \code{paragraph_id} columns.}
}
\value{
A tibble with lines in column \code{text}.
}
\description{
Takes an annotation-style dataframe and concatenates tokens
back into lines of text
}
\examples{
## Not run
collapse_annotation(df)

}
