% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_up.R
\name{cut_up}
\alias{cut_up}
\title{'Cut-up' a hymn (or something else)}
\usage{
cut_up(df, ref_id = 1)
}
\arguments{
\item{df}{A tibble based on output from \code{udpipe::udpipe_annotate()}
which is assumed to also contain a \code{wovel} column.}

\item{ref_id}{An integer number; this is the text that
will get new words.}
}
\value{
A tibble with new column \code{token_new} with a suggested
replacement for column \code{token}.
}
\description{
Using the cut-up technique popularized by William S. Burroughs,
\code{cut_up()} replaces each word in the reference document with
a random sample of words with same number of wovels
and same part-of-speach-tag
}
\examples{
## Not run
cut_up(hymns)

}
