% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rhymes.R
\name{get_rhymes}
\alias{get_rhymes}
\title{Find possible rhymes for a given word}
\usage{
get_rhymes(
  word,
  pron,
  token = token,
  stress_vowel = stress_vowel,
  remainder = remainder
)
}
\arguments{
\item{word}{character. Word to find rhymes for}

\item{pron}{A tibble assumed to contain a \code{token},
a \code{stress_vowel} and a \code{remainder} column, all
coded in then phonetic alphabet
\href{https://www.phon.ucl.ac.uk/home/sampa/index.html}{sampa}.}

\item{token}{character. Name of column containing token (word).}

\item{stress_vowel}{character. Name of column containing
pronounciation for the stressed vowel. Used for finding rhymes.}

\item{remainder}{character. Name of column containing
pronounciation for the remainer of the token, i.e. everything
\emph{after} the stressed vowel. Used for finding rhymes.}
}
\value{
A charactor vector with possible rhyming words.
}
\description{
Given a pronounciation dictionary with columns for
tokens (words) and pronounciations, this function
returns a vector of possible rhymes.
}
\examples{
pr <- data.frame(token = c("appellere", "akkumulere", "bygning"),
sampa = c("A$pE$\"le:?$6", "a$ku$mu$\"le:?$6", "\"by$gneN"),
rhyme_part = c("\"le:?$6", "\"le:?$6", "\"by$gneN"),
stress_vowel = c("e:?", "e:?", "y"),
remainder = c("$6", "$6", "$gneN"))

get_rhymes(word = "appellere", pron = pr)

}
