% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rhymes.R
\name{new_rhymes}
\alias{new_rhymes}
\title{Find rhymes for a hymn (or something else)}
\usage{
new_rhymes(hymn, df)
}
\arguments{
\item{hymn}{A tokenized hymn that needs new rhymes, possible output
from \code{salmer::cut_up()} in the same format as \code{annotated_hymns}}

\item{df}{A tibble based on output from \code{udpipe::udpipe_annotate()}}
}
\value{
A tibble with new column \code{token_new} with a suggested
replacement for column \code{token}.
}
\description{
Similar to \code{cut_up()}, \code{new_rhymes()} takes
a piece of tokenized text input, and replaces words in
rhyme scheme with words that rhyme.
}
