% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pronounciations}
\alias{pronounciations}
\title{Pronunciation Lexicon for Danish}
\format{
A tibble with 226,238 rows and 5 variables:
\describe{
\item{token}{word in Danish (mostly)}
\item{sampa}{Pronounciation in sampa phonetic alphabet}
\item{rhyme_part}{The part of pronounciation relevant for rhymes: Stressed syllable and everyhing after}
\item{stress_vowel}{Stressed syllable starting with first vowel}
\item{remainder}{Everything in the sampa string field after stressed syllable}
}
}
\usage{
pronounciations
}
\description{
A collection of Danish words with pronounciations available at
\href{https://sprogteknologi.dk/dataset/nst-lexical-database-for-danish}{sprogteknologi.dk}
or from
\href{https://www.nb.no/sprakbanken/ressurskatalog/oai-nb-no-sbr-26}{den norske sprogbank i Nationalbiblioteket}
where documentation is also available.
Each row corresponds to one Danish word with phonetic annotation using
\href{https://www.phon.ucl.ac.uk/home/sampa/index.html}{sampa}
-- a computer readable phonetic alphabet.
}
\keyword{datasets}
