% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhyme_scheme.R
\name{rhyme_scheme}
\alias{rhyme_scheme}
\title{Get rhyme scheme for a given hymn}
\usage{
rhyme_scheme(ref_id = 1, df, pron)
}
\arguments{
\item{ref_id}{An integer number, refering to a hymn number}

\item{df}{tibble based on output from \code{udpipe::udpipe_annotate()}
similar to the \code{annotated_hymns} dataset.}

\item{pron}{A tibble assumed to contain a \code{token},
a \code{stress_vowel} and a \code{remainder} column, all
coded in the phonetic alphabet. Passed on to \code{get_rhymes()}.}
}
\value{
a tibble with rhyme structure in new column, \code{scheme}. The
number, x, in that column indicates that the word must rhyme with the x-
previous word. I.e "2" means that this must rhyme with the word two lines
above.
}
\description{
Get rhyme scheme for a given hymn
}
