% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_annotation.R
\name{collapse_annotation}
\alias{collapse_annotation}
\title{'Collapses' a annotation back into text}
\usage{
collapse_annotation(
  df,
  token = token,
  doc_id = doc_id,
  line_id = line_id,
  text = text
)
}
\arguments{
\item{df}{A tibble based on output from \code{udpipe::udpipe_annotate()}}

\item{token}{Name of column containing tokens to be collapsed}

\item{doc_id}{Name of column containing document ids to group by}

\item{line_id}{Name of column containing line ids to group by}

\item{text}{Name of column to contain collapsed text}
}
\value{
A tibble with lines in column \code{text}.
}
\description{
Takes an annotation-style dataframe and concatenates tokens
back into lines of text
}
\examples{
## Not run
collapse_annotation(annotated_hymns)

}
