% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rhymes.R
\name{new_rhymes}
\alias{new_rhymes}
\title{'Cut-up' a hymn (or something else)}
\usage{
new_rhymes(hymn, df = annotated_hymns)
}
\arguments{
\item{hymn}{A tokenized hymn that needs new rhymes, possible output
from \code{salmer::cut_up()} in the same format as \code{annotated_hymns}}

\item{df}{A tibble based on output from \code{udpipe::udpipe_annotate()}}
}
\value{
A tibble with new column \code{token_new} with a suggested
replacement for column \code{token}.
}
\description{
Using the cut-up technique popularized by William S. Burroughs,
\code{cut_up()} replaces each word in the reference document with
a random sample of words with same number of wovels
and same part-of-speach-tag.
}
\examples{
## Not run
hymn <- annotated_hymns \%>\% dplyr::filter(doc_id == 1)
new_rhymes(hymn)

}
