% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhyme_scheme.R
\name{rhyme_scheme}
\alias{rhyme_scheme}
\title{Get rhyme scheme for a given hymn}
\usage{
rhyme_scheme(ref_id = 1, df)
}
\arguments{
\item{df}{tibble based on output from \code{udpipe::udpipe_annotate()}
similar to the \code{annotated_hymns} dataset.}
}
\value{
a tibble with rhyme structure in new column, \code{scheme}. The
number, x, in that column indicates that the word must rhyme with the x-
previous word. I.e "2" means that this must rhyme with the word two lines
above.
}
\description{
Get rhyme scheme for a given hymn
}
\examples{
## Not run
cut_up(annotated_hymns)

}
